+++ src/game/components/PhysicsBody.ts
@@ -13,14 +13,14 @@ import Vector3BufferView from "../../ecs/utils/Vector3BufferView";
 
 // TODO: optimize with ArrayBuffers
 class PhysicsBody extends Component {
-  private _values: Int32Array;
+  private _values: Float32Array;
   linearVelocity: Vector3BufferView;
   angularVelocity: Vector3BufferView;
 
   // TODO: ...
   constructor(entityId: EntityId) {
     super(entityId);
-    this._values = new Int32Array(6);
+    this._values = new Float32Array(6);
     this.linearVelocity = new Vector3BufferView(this._values);
     this.angularVelocity = new Vector3BufferView(this._values, 3 * 4);
   }
